refactor: remove BaseFormField
