+++ lib/ui/widgets/forms/form_date_field.dart
@@ -2,36 +2,42 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:mona/data/model/date.dart';
 import 'package:mona/l10n/build_context_extensions.dart';
-import 'package:mona/ui/widgets/forms/base_form_field.dart';
 
-class FormDateField extends BaseFormField {
+class FormDateField extends StatelessWidget {
   final Date date;
   final Date? selectedDate;
   final ValueChanged<Date> onChanged;
+  final String label;
+  final String? errorText;
 
-  FormDateField({
+  const FormDateField({
+    super.key,
     required this.date,
     required this.onChanged,
     this.selectedDate,
-    required super.label,
-    super.errorText,
+    required this.label,
+    this.errorText,
   });
 
   @override
-  Widget buildField(BuildContext context) {
-    return TextField(
-      controller: TextEditingController(
-          text: date.format(DateFormat.yMMMd(context.languageTag))),
-      keyboardType: TextInputType.datetime,
-      decoration: InputDecoration(
-        border: OutlineInputBorder(),
-        labelText: label,
-        errorText: errorText,
-        suffixIcon:
-            errorText != null ? Icon(Icons.error) : Icon(Icons.calendar_today),
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 8.0),
+      child: TextField(
+        controller: TextEditingController(
+            text: date.format(DateFormat.yMMMd(context.languageTag))),
+        keyboardType: TextInputType.datetime,
+        decoration: InputDecoration(
+          border: OutlineInputBorder(),
+          labelText: label,
+          errorText: errorText,
+          suffixIcon: errorText != null
+              ? Icon(Icons.error)
+              : Icon(Icons.calendar_today),
+        ),
+        readOnly: true,
+        onTap: () => _selectDate(context),
       ),
-      readOnly: true,
-      onTap: () => _selectDate(context),
     );
   }
