+++ lib/ui/widgets/forms/form_datetime_field.dart
@@ -2,63 +2,67 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:mona/l10n/build_context_extensions.dart';
 import 'package:mona/ui/constants/dimensions.dart';
-import 'package:mona/ui/widgets/forms/base_form_field.dart';
 
-class FormDateTimeField extends BaseFormField {
+class FormDateTimeField extends StatelessWidget {
   final DateTime datetime;
   final DateTime? selectedDatetime;
   final ValueChanged<DateTime> onChanged;
+  final String label;
+  final String? errorText;
 
   FormDateTimeField({
     required this.datetime,
     required this.onChanged,
     this.selectedDatetime,
-    required super.label,
-    super.errorText,
+    required this.label,
+    this.errorText,
   });
 
   @override
-  Widget buildField(BuildContext context) {
+  Widget build(BuildContext context) {
     final String locale = context.languageTag;
     final DateFormat timeFormat = MediaQuery.of(context).alwaysUse24HourFormat
         ? DateFormat.Hm(locale)
         : DateFormat.jm(locale);
 
-    return Row(
-      children: <Widget>[
-        Expanded(
-          child: TextField(
-            controller: TextEditingController(
-                text: DateFormat.yMMMd(locale).format(datetime)),
-            decoration: InputDecoration(
-              border: OutlineInputBorder(),
-              labelText: label,
-              errorText: errorText,
-              suffixIcon: errorText != null
-                  ? Icon(Icons.error)
-                  : Icon(Icons.calendar_today),
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 8.0),
+      child: Row(
+        children: <Widget>[
+          Expanded(
+            child: TextField(
+              controller: TextEditingController(
+                  text: DateFormat.yMMMd(locale).format(datetime)),
+              decoration: InputDecoration(
+                border: OutlineInputBorder(),
+                labelText: label,
+                errorText: errorText,
+                suffixIcon: errorText != null
+                    ? Icon(Icons.error)
+                    : Icon(Icons.calendar_today),
+              ),
+              readOnly: true,
+              onTap: () => _selectDate(context),
             ),
-            readOnly: true,
-            onTap: () => _selectDate(context),
           ),
-        ),
-        SizedBox(width: borderPadding),
-        IntrinsicWidth(
-          child: TextField(
-            controller:
-                TextEditingController(text: timeFormat.format(datetime)),
-            decoration: InputDecoration(
-              border: OutlineInputBorder(),
-              errorText: errorText,
-              suffixIcon: errorText != null
-                  ? Icon(Icons.error)
-                  : Icon(Icons.access_time),
+          SizedBox(width: borderPadding),
+          IntrinsicWidth(
+            child: TextField(
+              controller:
+                  TextEditingController(text: timeFormat.format(datetime)),
+              decoration: InputDecoration(
+                border: OutlineInputBorder(),
+                errorText: errorText,
+                suffixIcon: errorText != null
+                    ? Icon(Icons.error)
+                    : Icon(Icons.access_time),
+              ),
+              readOnly: true,
+              onTap: () => _selectTime(context),
             ),
-            readOnly: true,
-            onTap: () => _selectTime(context),
           ),
-        ),
-      ],
+        ],
+      ),
     );
   }
