+++ lib/ui/widgets/forms/form_text_field.dart
@@ -1,49 +1,56 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:mona/ui/widgets/forms/base_form_field.dart';
 
-class FormTextField extends BaseFormField {
+class FormTextField extends StatelessWidget {
   final TextEditingController controller;
   final TextInputType inputType;
   final VoidCallback onChanged;
   final bool readonly;
   final bool multiline;
+  final String label;
+  final String? suffixText;
+  final String? errorText;
+  final String? regexFormatter;
 
-  FormTextField({
+  const FormTextField({
+    super.key,
     required this.controller,
     required this.inputType,
     required this.onChanged,
-    required super.label,
-    super.suffixText,
-    super.errorText,
-    super.regexFormatter,
+    required this.label,
+    this.suffixText,
+    this.errorText,
+    this.regexFormatter,
     this.readonly = false,
     this.multiline = false,
   });
 
   @override
-  Widget buildField(BuildContext context) {
-    return TextField(
-      controller: controller,
-      keyboardType: multiline ? TextInputType.multiline : inputType,
-      inputFormatters: regexFormatter != null
-          ? <TextInputFormatter>[
-              FilteringTextInputFormatter.allow(
-                RegExp(regexFormatter!),
-              )
-            ]
-          : null,
-      decoration: InputDecoration(
-        border: OutlineInputBorder(),
-        labelText: label,
-        suffixText: suffixText,
-        errorText: errorText,
-        suffixIcon: errorText != null ? Icon(Icons.error) : null,
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 8.0),
+      child: TextField(
+        controller: controller,
+        keyboardType: multiline ? TextInputType.multiline : inputType,
+        inputFormatters: regexFormatter != null
+            ? <TextInputFormatter>[
+                FilteringTextInputFormatter.allow(
+                  RegExp(regexFormatter!),
+                )
+              ]
+            : null,
+        decoration: InputDecoration(
+          border: OutlineInputBorder(),
+          labelText: label,
+          suffixText: suffixText,
+          errorText: errorText,
+          suffixIcon: errorText != null ? Icon(Icons.error) : null,
+        ),
+        onChanged: (value) => onChanged(),
+        readOnly: readonly,
+        maxLines: multiline ? null : 1,
+        minLines: multiline ? 3 : null,
       ),
-      onChanged: (value) => onChanged(),
-      readOnly: readonly,
-      maxLines: multiline ? null : 1,
-      minLines: multiline ? 3 : null,
     );
   }
 }
